feat: ensure product quantity can be decreased

+++ src/actions/decrease-cart-products-quantity/index.ts
@@ -8,12 +8,14 @@ import { cartItemTable, cartTable } from "@/db/schema";
 import { auth } from "@/lib/auth";
 
 import {
-  RemoveProductFromCartSchema,
-  removeProductFromCartSchema,
+  DecreaseCartProductsQuantitySchema,
+  decreaseCartProductsQuantitySchema,
 } from "./schema";
 
-export const removeProductToCart = async (data: RemoveProductFromCartSchema) => {
-  removeProductFromCartSchema.parse(data);
+export const decreaseCartProductQuantity = async (
+  data: DecreaseCartProductsQuantitySchema,
+) => {
+  decreaseCartProductsQuantitySchema.parse(data);
   const session = await auth.api.getSession({
     headers: await headers(),
   });
@@ -43,13 +45,21 @@ export const removeProductToCart = async (data: RemoveProductFromCartSchema) => 
       cart: true,
     },
   });
-    if (!cartItem) {
-      throw new Error("Cart item not found");
-    }
+  if (!cartItem) {
+    throw new Error("Cart item not found");
+  }
   const cartDoesNotBelongToUser = cartItem.cart.userId !== session.user.id;
   if (cartDoesNotBelongToUser) {
     throw new Error("Unauthorized");
   }
-
-  await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+  if (cartItem.quantity === 1) {
+    await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+    return;
+  }
+  await db
+    .update(cartItemTable)
+    .set({
+      quantity: cartItem.quantity - 1,
+    })
+    .where(eq(cartItemTable.id, cartItem.id));
 };
